refactor: remove CreateDesignContainer component
